requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-login==0.6.3
werkzeug==3.1.3
bcrypt>=4.0.0
pillow>=10.0.0 